+++ Neolution.DotNet.Console.SampleAsync/Commands/GuidGen/GuidGenCommand.cs
@@ -8,8 +8,8 @@
     /// <summary>
     /// Prints a randomly generated GUID in the console window
     /// </summary>
-    /// <seealso cref="IConsoleAppCommand{TOptions}" />
-    public class GuidGenCommand : IConsoleAppCommand<GuidGenOptions>
+    /// <seealso cref="IDotNetConsoleCommand{TOptions}" />
+    public class GuidGenCommand : IDotNetConsoleCommand<GuidGenOptions>
     {
         /// <summary>
         /// The logger