+++ Neolution.DotNet.Console.UnitTests/GrammarTests/Fakes/EchoCommand.cs
@@ -1,14 +1,14 @@
-﻿namespace Neolution.DotNet.Console.UnitTests.ConsoleAppGrammar.Fakes
+﻿namespace Neolution.DotNet.Console.UnitTests.GrammarTests.Fakes
 {
     using System.Threading.Tasks;
     using Neolution.DotNet.Console.Abstractions;
     using Neolution.DotNet.Console.UnitTests.Common.Spies;
 
     /// <summary>
-    /// The command for the default verb.
+    /// Fake command that logs the value passed via command line.
     /// </summary>
-    /// <seealso cref="IConsoleAppCommand{DefaultVerbOptions}" />
-    public class DefaultCommand : IConsoleAppCommand<DefaultOptions>
+    /// <seealso cref="IDotNetConsoleCommand{TOptions}" />
+    public class EchoCommand : IDotNetConsoleCommand<EchoOptions>
     {
         /// <summary>
         /// The logger
@@ -16,18 +16,18 @@
         private readonly IUnitTestLogger logger;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="DefaultCommand"/> class.
+        /// Initializes a new instance of the <see cref="EchoCommand"/> class.
         /// </summary>
         /// <param name="logger">The logger.</param>
-        public DefaultCommand(IUnitTestLogger logger)
+        public EchoCommand(IUnitTestLogger logger)
         {
             this.logger = logger;
         }
 
         /// <inheritdoc />
-        public async Task RunAsync(DefaultOptions options)
+        public async Task RunAsync(EchoOptions options)
         {
-            this.logger.Log(options);
+            this.logger.Log("options", options);
             await Task.CompletedTask;
         }
     }