+++ Neolution.DotNet.Console.UnitTests/RunTests/DotNetConsoleRunTests.cs
@@ -1,0 +1,62 @@
+﻿namespace Neolution.DotNet.Console.UnitTests.RunTests
+{
+    using System;
+    using System.Reflection;
+    using System.Threading.Tasks;
+    using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Hosting;
+    using Neolution.DotNet.Console.UnitTests.Common.Fakes;
+    using Neolution.DotNet.Console.UnitTests.Common.Spies;
+    using Neolution.DotNet.Console.UnitTests.Common.Stubs;
+    using Shouldly;
+    using Xunit;
+
+    /// <summary>
+    /// Run tests for the <see cref="DotNetConsole"/>.
+    /// </summary>
+    [Collection("Non-Parallel Tests")]
+    public class DotNetConsoleRunTests
+    {
+        /// <summary>
+        /// Given an environment variable with an environment name, when running the console application, then the environment name of the app should match.
+        /// </summary>
+        /// <param name="environmentName">Name of the environment.</param>
+        /// <returns>The <see cref="Task"/>.</returns>
+        [Theory]
+        [InlineData("Development")]
+        [InlineData("Staging")]
+        [InlineData("Production")]
+        public async Task GivenEnvironmentNameEnvironmentVariable_WhenRunningConsoleApp_ThenEnvironmentNameShouldMatch(string environmentName)
+        {
+            // Arrange
+            Environment.SetEnvironmentVariable("DOTNET_ENVIRONMENT", environmentName);
+
+            const string commandLineString = "default";
+            var builder = DotNetConsole.CreateBuilderWithReference(Assembly.GetAssembly(typeof(DefaultCommand))!, commandLineString.Split(" "));
+
+            if (builder.Environment.IsDevelopment())
+            {
+                // In Development, all services must be added because validation makes sure that each registration can be constructed by DI.
+                builder.Services.AddTransient<ITransientServiceStub, TransientServiceStub>();
+                builder.Services.AddScoped<IScopedServiceStub, ScopedServiceStub>();
+                builder.Services.AddSingleton<ISingletonServiceStub, SingletonServiceStub>();
+            }
+
+            var logger = new UnitTestLogger();
+            builder.Services.AddSingleton(typeof(IUnitTestLogger), logger);
+
+            var console = builder.Build();
+
+            // Act
+            await console.RunAsync();
+
+            // Assert
+            builder.Environment.EnvironmentName.ShouldBe(environmentName);
+            var environment = (IHostEnvironment)logger.LoggedObjects["environment"];
+            environment.EnvironmentName.ShouldBe(environmentName);
+
+            // Cleanup
+            Environment.SetEnvironmentVariable("DOTNET_ENVIRONMENT", null);
+        }
+    }
+}