+++ Neolution.DotNet.Console.UnitTests/RunTests/DotNetConsoleStartupTests.cs
@@ -1,17 +1,18 @@
-﻿namespace Neolution.DotNet.Console.UnitTests.ConsoleAppStartup
+﻿namespace Neolution.DotNet.Console.UnitTests.RunTests
 {
     using System.Reflection;
     using Microsoft.Extensions.DependencyInjection;
     using Neolution.DotNet.Console.UnitTests.Common;
+    using Neolution.DotNet.Console.UnitTests.Common.Fakes;
     using Neolution.DotNet.Console.UnitTests.Common.Stubs;
-    using Neolution.DotNet.Console.UnitTests.ConsoleAppGrammar.Fakes;
     using Shouldly;
     using Xunit;
 
     /// <summary>
     /// Tests for the Console App Startup class.
     /// </summary>
-    public class ConsoleAppStartupTests
+    [Collection("Production Environment Tests")]
+    public class DotNetConsoleStartupTests
     {
         /// <summary>
         /// Given configured services with various (all possible) service lifetimes, when running the console application, then should it not throw.